Farm-out timeline data
Moving it into separate files/folders will make it easier to manage once all backpacking dates have been added.
