+++ src/lib/timelineData/greeceData.js
@@ -1,0 +1,13 @@
+import { countries, places } from '@/lib/placeNames'
+import { dates, arrival, departure } from '@/lib/timelineData/variables'
+
+export const greeceData = {
+  [places[countries.greece].athens]: [
+    {
+      [dates]: {
+        [arrival]: '2023-08-10',
+        [departure]: '2023-10-26',
+      },
+    },
+  ],
+}
