+++ src/lib/timelineData/thailandData/bangkokData.js
@@ -1,0 +1,34 @@
+import { dates, arrival, departure } from '@/lib/timelineData/variables'
+
+export const bangkokData = [
+  {
+    [dates]: {
+      [arrival]: '2022-10-07',
+      [departure]: '2022-11-03',
+    },
+  },
+  {
+    [dates]: {
+      [arrival]: '2022-12-01',
+      [departure]: '2022-12-16',
+    },
+  },
+  {
+    [dates]: {
+      [arrival]: '2023-03-11',
+      [departure]: '2023-03-23',
+    },
+  },
+  {
+    [dates]: {
+      [arrival]: '2023-04-06',
+      [departure]: '2023-04-20',
+    },
+  },
+  {
+    [dates]: {
+      [arrival]: '2023-05-11',
+      [departure]: '2023-05-24',
+    },
+  },
+]
