+++ src/lib/timelineData/variables.js
@@ -1,0 +1,5 @@
+export const dates = 'dates'
+export const arrival = 'arrival'
+export const departure = 'departure'
+export const country = 'country'
+export const place = 'place'
